refactor: login component
