+++ src/slices/user.ts
@@ -18,34 +18,6 @@ const initialState: UserDetail = {
 
 const host = checkEnvironment();
 
-export const loginUser = createAsyncThunk('user/login', async (obj, { getState }) => {
-  const { user } = getState() as { user: UserDetail };
-
-  const data = {
-    email: user.email,
-    password: user.password
-  };
-
-  const url = `${host}/api/login`;
-
-  const response = await fetch(url, {
-    method: 'POST',
-    mode: 'cors',
-    cache: 'no-cache',
-    credentials: 'same-origin',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    redirect: 'follow',
-    referrerPolicy: 'no-referrer',
-    body: JSON.stringify(data)
-  });
-
-  const dataInJson = await response.json();
-
-  return dataInJson;
-});
-
 export const fetchUser = createAsyncThunk('users/fetchUser', async (obj, { getState }) => {
   const { user } = getState() as { user: UserDetail };
 
@@ -65,25 +37,6 @@ export const userSlice = createSlice({
     resetUserData: () => initialState
   },
   extraReducers: {
-    [loginUser.pending.toString()]: (state) => {
-      state.status = 'pending';
-      state.isFetching = true;
-    },
-    [loginUser.fulfilled.toString()]: (state, { payload }) => {
-      state.status = 'success';
-      state.id = payload && payload.id;
-      state.password = '';
-      state.confirmPassword = '';
-      state.error = (payload && payload.error) || '';
-      state.isFetching = false;
-      state.message = payload && payload.message;
-    },
-    [loginUser.rejected.toString()]: (state, { payload }) => {
-      state.status = 'failed';
-      state.isFetching = false;
-      state.error = (payload && payload.error) || '';
-      state.message = payload && payload.message;
-    },
     [fetchUser.pending.toString()]: (state) => {
       state.status = 'pending';
     },
